api/requirements.txt
Django>=4.0,<5.0  # Or your preferred Django version
psycopg2-binary   # PostgreSQL adapter
gunicorn          # WSGI server
dj-database-url   # For easier DB URL parsing (optional but good practice)
dj-rest-auth[with_social]
dj_database_url
djangorestframework
drf-yasg
google-cloud-storage
google-cloud-vision 
googlemaps 
Pillow
whitenoise 